Move all extensions to its own module
