+++ chat/__init__.py
@@ -1,42 +1,35 @@
 import os
-from celery import Celery
 from dotenv import load_dotenv
 from flask import Flask, g, session
-from flask_marshmallow import Marshmallow
-from flask_moment import Moment
-from flask_socketio import SocketIO
-from flask_sqlalchemy import SQLAlchemy
 
-from config.app import Config
-
-# Load environment variables from .env file
-load_dotenv()
-
-celery = Celery(__name__, broker=Config.CELERY_BROKER_URL)
-
-db = SQLAlchemy()
-ma = Marshmallow()
-moment = Moment()
-socketio = SocketIO()
+from . import auth, chat
+from .extensions import (
+    celery,
+    db,
+    ma,
+    moment,
+    socketio
+)
+from .models import User
 
 def create_app():
+    # Load environment variables from .env file
+    load_dotenv()
+
     app = Flask(__name__)
     app.config.from_object(os.environ['APP_SETTINGS'])
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
-    celery.conf.update(app.config)
-
-    db.init_app(app)
-    ma.init_app(app)
-    moment.init_app(app)
-    socketio.init_app(app)
-
     with app.app_context():
-        # Models
-        from .models import User
+        celery.conf.update(app.config)
 
-        # Blueprints
-        from . import auth, chat
+        # Init extensions
+        db.init_app(app)
+        ma.init_app(app)
+        moment.init_app(app)
+        socketio.init_app(app)
+
+        # Register blueprints
         app.register_blueprint(auth.bp)
         app.register_blueprint(chat.bp)
 
@@ -53,4 +46,4 @@ def create_app():
         # Create db models
         db.create_all()
 
-        return app
+    return app
